update ble spam and move to base pack

+++ base_pack/apple_ble_spam/lib/continuity/continuity.h
@@ -4,9 +4,11 @@
 #include <stdlib.h>
 
 // Hacked together by @Willy-JL
+// Custom adv logic by @Willy-JL and @xMasterX
+// Extensive testing and research on behavior and parameters by @Willy-JL and @ECTO-1A
 // Structures docs and Nearby Action IDs from https://github.com/furiousMAC/continuity/
 // Proximity Pair IDs from https://github.com/ECTO-1A/AppleJuice/
-// Custom adv logic and Airtag ID from https://techryptic.github.io/2023/09/01/Annoying-Apple-Fans/
+// Airtag ID from https://techryptic.github.io/2023/09/01/Annoying-Apple-Fans/
 
 typedef enum {
     ContinuityTypeAirDrop = 0x05,
@@ -44,6 +46,6 @@ typedef struct {
 
 const char* continuity_get_type_name(ContinuityType type);
 
-size_t continuity_get_packet_size(ContinuityType type);
+uint8_t continuity_get_packet_size(ContinuityType type);
 
 void continuity_generate_packet(const ContinuityMsg* msg, uint8_t* packet);
